made the thing move better

+++ src/entity/entity.hpp
@@ -21,6 +21,6 @@ class DrawComponent {
 class Entity {
  public:
   std::vector<DrawComponent *> textures;
-  int x = 0;
-  int y = 0;
+  double x = 0;
+  double y = 0;
 };
